Combine PasswordInput and PasswordInputStateful
